getcwd and formatting
