+++ c-scape/src/io/mod.rs
@@ -1,6 +1,6 @@
 #[cfg(not(target_os = "wasi"))]
 mod dup;
 mod isatty;
+mod pipe;
 mod read;
 mod write;
-mod pipe;
