+++ c-scape/src/io/pipe.rs
